Read RoboCupRescue Simulation map data from GML file

This function reads and processes map data for RoboCupRescue Simulation from a GML file. It extracts nodes, edges, buildings and roads, and organizes them into a list.

Usage

read_rrs_map(gml, scale_data = FALSE)

Arguments

gml

Path to the GML file.

scale_data

Logical. If TRUE, coordinates are scaled up by a factor of 1000 to match the simulation environment, and adjusted such that the minimum x and y values are 0. (Default: TRUE)

Value

A list of sf objects: nodes, edges, buildings, and roads. Also includes a sub-list raw_parsed_data containing the intermediate data frames from XML parsing.

Examples

gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map_data <- read_rrs_map(gml)
map_data
#> $node_sf
#> Simple feature collection with 315 features and 1 field
#> Attribute-geometry relationship: constant (1)
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -30 ymin: -20 xmax: 135 ymax: 121
#> CRS:           NA
#> First 10 features:
#>    id            geometry
#> 1   0 POINT (83.16 43.16)
#> 2   1 POINT (76.84 36.84)
#> 3   2  POINT (76.84 3.16)
#> 4   3  POINT (83.16 3.16)
#> 5   4   POINT (3.16 3.16)
#> 6   5 POINT (-3.16 -3.16)
#> 7   6    POINT (80 -3.16)
#> 8   7       POINT (67 43)
#> 9   8 POINT (73.16 36.84)
#> 10  9    POINT (40 -3.16)
#> 
#> $edge_sf
#> Simple feature collection with 413 features and 1 field
#> Attribute-geometry relationship: constant (1)
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -30 ymin: -20 xmax: 135 ymax: 121
#> CRS:           NA
#> First 10 features:
#>     id                       geometry
#> 1  103 LINESTRING (83.16 43.16, 76...
#> 2  110 LINESTRING (80 -3.16, 76.84...
#> 3  108 LINESTRING (83.16 3.16, 83....
#> 4  109 LINESTRING (3.16 3.16, -3.1...
#> 5  106 LINESTRING (76.84 36.84, 76...
#> 6  107 LINESTRING (76.84 3.16, 83....
#> 7  104 LINESTRING (103.16 2.24, 10...
#> 8  119 LINESTRING (40 -3.16, 80 -3...
#> 9  117 LINESTRING (81.31 103.16, 4...
#> 10 116 LINESTRING (73.16 36.84, 76...
#> 
#> $building_sf
#> Simple feature collection with 37 features and 1 field
#> Attribute-geometry relationship: constant (1)
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -30 ymin: -20 xmax: 135 ymax: 121
#> CRS:           NA
#> First 10 features:
#>      id                       geometry
#> 247 247 POLYGON ((-20 25, -30 30, -...
#> 248 248 POLYGON ((10 -20, 10 -10, 1...
#> 249 249 POLYGON ((-30 40, -30 70, -...
#> 250 250 POLYGON ((92 95, 90 95, 90 ...
#> 251 251 POLYGON ((40 -20, 30 -20, 3...
#> 253 253 POLYGON ((42.16 103.16, 40 ...
#> 254 254 POLYGON ((135 50, 125 50, 1...
#> 255 255 POLYGON ((105 5, 108 -8, 98...
#> 298 298 POLYGON ((-3.193 102.15, -1...
#> 905 905 POLYGON ((18 76, 22 73, 22 ...
#> 
#> $road_sf
#> Simple feature collection with 58 features and 1 field
#> Attribute-geometry relationship: constant (1)
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -11.2 ymin: -10 xmax: 103.16 ymax: 103.16
#> CRS:           NA
#> First 10 features:
#>      id                       geometry
#> 256 256 POLYGON ((3 77, 3 84, 18 84...
#> 257 257 POLYGON ((61 77, 59 83, 84....
#> 258 258 POLYGON ((73.16 26.84, 67 3...
#> 259 259 POLYGON ((92.24 92.24, 84.1...
#> 262 262 POLYGON ((103.16 81.31, 96....
#> 263 263 POLYGON ((53 62, 45 68, 42....
#> 266 266 POLYGON ((76.84 36.84, 83.1...
#> 268 268 POLYGON ((47 58, 53 62, 53 ...
#> 269 269 POLYGON ((73.16 36.84, 67 4...
#> 270 270 POLYGON ((49 77, 42 83, 59 ...
#> 
#> $raw_parsed_data
#> $raw_parsed_data$nodes
#>      id       x       y
#> 1     0  83.160  43.160
#> 2     1  76.840  36.840
#> 3     2  76.840   3.160
#> 4     3  83.160   3.160
#> 5     4   3.160   3.160
#> 6     5  -3.160  -3.160
#> 7     6  80.000  -3.160
#> 8     7  67.000  43.000
#> 9     8  73.160  36.840
#> 10    9  40.000  -3.160
#> 11   10  51.410  12.830
#> 12   11  48.590   7.170
#> 13   13  40.000  11.460
#> 14   14 103.160  70.000
#> 15   15  53.000  62.000
#> 16   17  49.000  77.000
#> 17   16  45.000  68.000
#> 18   19  42.000  83.000
#> 19   18 103.160   2.240
#> 20   21  -3.160  80.000
#> 21   20   3.000  77.000
#> 22   23  53.000  34.000
#> 23   22  47.000  27.000
#> 24   25  73.160  26.840
#> 25   24  67.000  34.000
#> 26   27 103.160  81.310
#> 27   29  96.840  78.690
#> 28   28  96.840   3.160
#> 29   31  61.000  77.000
#> 30   30  89.150  86.380
#> 31   34  81.310 103.160
#> 32   32  59.000  83.000
#> 33   33  78.690  96.840
#> 34   38  31.410   7.170
#> 35   39   3.000  97.000
#> 36   36  84.150  91.380
#> 37   37  28.590  12.830
#> 38   42  49.090  23.030
#> 39   43  40.000  20.300
#> 40   40  92.240  92.240
#> 41   41   3.000  84.000
#> 42   46  30.910  23.030
#> 43   44  50.910  16.970
#> 44   45  29.090  16.970
#> 45   51 105.000   5.000
#> 46   50  21.000  84.000
#> 47   49  19.040  76.840
#> 48   48  47.000  58.000
#> 49   55 -20.000  50.000
#> 50   54  95.000 -10.000
#> 51   53  98.000 -20.000
#> 52   52 108.000  -8.000
#> 53   59 -30.000  40.000
#> 54   58 -10.000  70.000
#> 55   57 -10.000  60.000
#> 56   56 -20.000  60.000
#> 57   63  10.000 -10.000
#> 58   62 -10.000  50.000
#> 59   61 -10.000  40.000
#> 60   60 -30.000  70.000
#> 61   68 -20.000  25.000
#> 62   69 -10.000  30.000
#> 63   70 -20.000  35.000
#> 64   71  30.000 -20.000
#> 65  609   4.000  58.000
#> 66   64  10.000 -20.000
#> 67   65  20.000 -20.000
#> 68  615   3.048  55.000
#> 69   66  20.000 -10.000
#> 70  612   3.041  58.000
#> 71   67 -30.000  30.000
#> 72   76  40.000 101.000
#> 73  618  23.080  74.118
#> 74   77  30.000 111.000
#> 75   78  40.000 121.000
#> 76   79  90.000  95.000
#> 77   72  30.000  -1.000
#> 78   73  40.000  -1.000
#> 79   74  40.000 -20.000
#> 80   75  50.000 111.000
#> 81  621  45.000  53.000
#> 82   85 -10.000 115.000
#> 83  627  47.000  57.000
#> 84   84 -10.000  95.000
#> 85   87 135.000  70.000
#> 86   86   0.000 105.000
#> 87  624  47.000  53.000
#> 88   81 100.000  90.000
#> 89   80  95.000  95.000
#> 90  630   7.000  96.000
#> 91   83 -20.000 105.000
#> 92   82  95.000  90.000
#> 93   93 100.000 105.000
#> 94   92 101.000  60.000
#> 95   95  99.560  -3.160
#> 96  633   7.000  96.992
#> 97   94  90.000 105.000
#> 98  639  21.000  85.000
#> 99   89 125.000  50.000
#> 100  88 101.000  70.000
#> 101  91 125.000  60.000
#> 102 636   4.000  96.998
#> 103  90 135.000  50.000
#> 104 102  37.840 103.100
#> 105 100 101.212  -0.664
#> 106 101  42.160 103.160
#> 107  98  33.950  15.510
#> 108  99 103.160  60.000
#> 109 582   3.156   5.000
#> 110  96  46.050  15.510
#> 111  97  30.000  -3.160
#> 112 585   5.000   3.160
#> 113 591   3.104  29.000
#> 114 594   3.098  32.000
#> 115 597   4.000  32.000
#> 116 600   4.000  29.000
#> 117 606   4.000  55.000
#> 118 687  95.000  74.000
#> 119 684  92.265  83.265
#> 120 681  93.765  81.765
#> 121 678  93.000  81.000
#> 122 675  91.500  82.500
#> 123 672  60.000  96.880
#> 124 702  96.840  68.000
#> 125 699  95.000  68.000
#> 126 696  96.840  74.000
#> 127 693  96.840  72.000
#> 128 690  95.000  72.000
#> 129 654  53.000  95.000
#> 130 648  41.000  83.000
#> 131 651  42.769  79.154
#> 132 645  21.000  90.000
#> 133 642  18.000  84.000
#> 134 669  63.000  96.873
#> 135 666  63.000  96.000
#> 136 660  49.000  96.903
#> 137 663  53.002  96.894
#> 138 657  49.000  95.000
#> 139 747  96.840  38.000
#> 140 744  96.840  40.000
#> 141 750  95.000  38.000
#> 142 738  96.840  44.000
#> 143 741  95.000  44.000
#> 144 762  96.840  28.000
#> 145 765  96.840  26.000
#> 146 753  95.000  31.000
#> 147 759  96.840  29.000
#> 148 756  96.840  31.000
#> 149 714  96.840  59.000
#> 150 717  96.840  58.000
#> 151 705  96.840  70.000
#> 152 708  95.000  61.000
#> 153 711  96.840  61.000
#> 154 729  96.840  49.000
#> 155 732  96.840  47.000
#> 156 735  96.840  46.000
#> 157 720  96.840  56.000
#> 158 723  95.000  56.000
#> 159 726  95.000  49.000
#> 160 822  76.999  43.099
#> 161 816  73.000  44.000
#> 162 819  72.999  43.059
#> 163 828  54.000  77.000
#> 164 831  51.000  77.000
#> 165 825  54.000  76.000
#> 166 287  -2.170 103.119
#> 167 286  -3.193 102.150
#> 168 804  63.000   7.000
#> 169 807  63.000   3.160
#> 170 801  68.000   7.000
#> 171 813  77.000  44.000
#> 172 810  68.000   3.160
#> 173 305  -3.160  46.000
#> 174 789  95.000   7.000
#> 175 786  95.000  14.000
#> 176 308  -3.160  44.000
#> 177 798  96.840   7.000
#> 178 315 -11.200  29.400
#> 179 795  96.840   9.000
#> 180 792  95.000   9.000
#> 181 318 -11.200  30.600
#> 182 774  96.840  19.000
#> 183 771  95.000  19.000
#> 184 768  95.000  26.000
#> 185 783  96.840  14.000
#> 186 780  96.840  16.000
#> 187 299 -10.000  44.000
#> 188 302 -10.000  46.000
#> 189 777  96.840  17.000
#> 190 881  38.000   3.160
#> 191 343  10.000  -3.160
#> 192 340  13.000  -3.160
#> 193 884  42.000   3.160
#> 194 337  13.000 -10.000
#> 195 887  42.184  10.369
#> 196 351  18.000  76.000
#> 197 350   4.000  76.000
#> 198 890  37.816  10.369
#> 199 324  -3.160  29.000
#> 200 321  -3.160  31.000
#> 201 331  92.000  95.000
#> 202 369  23.000  10.000
#> 203 368  20.000   5.000
#> 204 370  23.000  17.000
#> 205 356   4.000  65.000
#> 206 357  22.000  53.000
#> 207 358  17.000  53.000
#> 208 359  17.000  48.000
#> 209 352  22.000  73.000
#> 210 353  22.000  68.000
#> 211 354   4.000  68.000
#> 212 355  22.000  65.000
#> 213 364  13.000  46.000
#> 214 365   4.000  17.000
#> 215 366   6.000   7.000
#> 216 367   8.000   5.000
#> 217 360   4.000  48.000
#> 218 361   4.000  46.000
#> 219 362   4.000  18.000
#> 220 363  13.000  18.000
#> 221 410  85.000  10.000
#> 222 411  95.000  10.000
#> 223 408  95.000   5.000
#> 224 409  85.000   5.000
#> 225 414  85.000  11.000
#> 226 415  85.000  16.000
#> 227 412  95.000  11.000
#> 228 413  95.000  16.000
#> 229 402  45.000  46.000
#> 230 403  24.000  59.000
#> 231 400  24.000  71.000
#> 232 401  45.000  57.000
#> 233 395  56.000   7.000
#> 234 394  74.000   7.000
#> 235 393  74.000  25.000
#> 236 392  56.000  25.000
#> 237 440  95.000  53.000
#> 238 441  85.000  53.000
#> 239 442  85.000  58.000
#> 240 443  95.000  58.000
#> 241 444  95.000  59.000
#> 242 445  85.000  59.000
#> 243 446  85.000  64.000
#> 244 447  95.000  64.000
#> 245 432  95.000  41.000
#> 246 433  85.000  41.000
#> 247 434  85.000  46.000
#> 248 435  95.000  46.000
#> 249 436  95.000  47.000
#> 250 437  85.000  47.000
#> 251 438  85.000  52.000
#> 252 439  95.000  52.000
#> 253 425  85.000  29.000
#> 254 424  95.000  29.000
#> 255 427  95.000  34.000
#> 256 426  85.000  34.000
#> 257 429  85.000  35.000
#> 258 428  95.000  35.000
#> 259 431  95.000  40.000
#> 260 430  85.000  40.000
#> 261 417  85.000  22.000
#> 262 416  85.000  17.000
#> 263 419  95.000  17.000
#> 264 418  95.000  22.000
#> 265 421  85.000  23.000
#> 266 420  95.000  23.000
#> 267 423  95.000  28.000
#> 268 422  85.000  28.000
#> 269 478  29.000  90.000
#> 270 479  29.000  96.000
#> 271 476  18.000  96.000
#> 272 477  18.000  90.000
#> 273 474  16.000  85.000
#> 274 475  16.000  96.000
#> 275 472   4.000  96.000
#> 276 473   4.000  85.000
#> 277 470  62.000  62.000
#> 278 471  55.000  62.000
#> 279 468  51.000  76.000
#> 280 469  61.000  76.000
#> 281 466  80.000  44.000
#> 282 467  80.000  55.000
#> 283 464  54.000  35.000
#> 284 465  66.000  44.000
#> 285 463  54.000  55.000
#> 286 462  66.000  55.000
#> 287 461  66.000  35.000
#> 288 460  95.000  79.000
#> 289 459  90.000  84.000
#> 290 458  85.000  84.000
#> 291 457  85.000  77.000
#> 292 456  95.000  77.000
#> 293 455  95.000  76.000
#> 294 454  85.000  76.000
#> 295 453  85.000  71.000
#> 296 452  95.000  71.000
#> 297 451  95.000  70.000
#> 298 450  85.000  70.000
#> 299 449  85.000  65.000
#> 300 448  95.000  65.000
#> 301 493  44.000  85.000
#> 302 492  44.000  95.000
#> 303 489  60.000  85.000
#> 304 488  60.000  96.000
#> 305 491  58.000  95.000
#> 306 490  58.000  85.000
#> 307 485  41.000  96.000
#> 308 963  30.000   3.160
#> 309 484  41.000  79.000
#> 310 487  78.000  96.000
#> 311 486  82.000  92.000
#> 312 481  29.000  85.000
#> 313 480  18.000  85.000
#> 314 483  33.000  79.000
#> 315 482  33.000  96.000
#> 
#> $raw_parsed_data$edges
#>      id start_node_id end_node_id
#> 1   103             0           1
#> 2   110             6           2
#> 3   108             3           0
#> 4   109             4           5
#> 5   106             1           2
#> 6   107             2           3
#> 7   104            18          99
#> 8   119             9           6
#> 9   117            34         101
#> 10  116             8           1
#> 11  115           101          75
#> 12  112             7           8
#> 13  127            51          18
#> 14  126            87          14
#> 15  123            14          88
#> 16  122             9          73
#> 17  121            11          10
#> 18  120            74           9
#> 19  136            23          22
#> 20  139            22          25
#> 21  138            25          24
#> 22  141            24          23
#> 23  140             3           6
#> 24  142            28           3
#> 25  129            14          27
#> 26  128            18         100
#> 27  131            17          19
#> 28  130            16          15
#> 29  133            15          17
#> 30  134            20          21
#> 31  153            41          39
#> 32  154            36          40
#> 33  157            21          41
#> 34  158            44          42
#> 35  159            42          43
#> 36  144            27          29
#> 37  145            32          31
#> 38  146            33          34
#> 39  147            31          30
#> 40  148            30          36
#> 41  149            36          32
#> 42  150            37          38
#> 43  171            41          20
#> 44  170            15          48
#> 45  169            23          15
#> 46  175            50          49
#> 47  173            48          16
#> 48  172            16          50
#> 49  163            46          45
#> 50  162             7          24
#> 51  161            33          36
#> 52  160            25           8
#> 53  167            32          19
#> 54  165            40          30
#> 55  164            43          46
#> 56  186            60          59
#> 57  187            59          61
#> 58  184            56          57
#> 59  185            52          51
#> 60  190            63          64
#> 61  191            64          65
#> 62  189            70          67
#> 63  178            54          53
#> 64  177            53          52
#> 65  182            62          55
#> 66  183            55          56
#> 67  180            20          49
#> 68  205            73          72
#> 69  207            89          90
#> 70  206            83          84
#> 71  201            86          85
#> 72  200            75          78
#> 73  203            80          82
#> 74  202            85          83
#> 75  197            71          74
#> 76  196            78          77
#> 77  199            65          66
#> 78  193            58          60
#> 79  192            57          58
#> 80  194            67          68
#> 81  220            96          44
#> 82  221            72          97
#> 83  222            97          71
#> 84  223            88          92
#> 85  216             6          95
#> 86  217            10          96
#> 87  212            82          81
#> 88  213            81          93
#> 89  208            90          87
#> 90  209            91          89
#> 91  210            94          79
#> 92  211            93          94
#> 93  237            79          34
#> 94  236            82          40
#> 95  235            27          82
#> 96  234            76         101
#> 97  232            95          54
#> 98  230           100          28
#> 99  229           100          95
#> 100 228            92          99
#> 101 227            99          91
#> 102 225            98          37
#> 103 244            77         102
#> 104 245            45          98
#> 105 240            40          79
#> 106 241           102          76
#> 107 304           302          62
#> 108 306            21         305
#> 109 309           305         308
#> 110 313           299         308
#> 111 312           302         305
#> 112 317           315          69
#> 113 316            68         315
#> 114 319            69         318
#> 115 288            86         287
#> 116 289           287         286
#> 117 290           286          84
#> 118 292            21         286
#> 119 293           102         287
#> 120 294            39         286
#> 121 300            61         299
#> 122 303           299         302
#> 123 342           340          97
#> 124 339           337          63
#> 125 338            66         337
#> 126 348           337         340
#> 127 347           343          63
#> 128 345           343         340
#> 129 344             5         343
#> 130 326           324           5
#> 131 325           321         324
#> 132 322           308         321
#> 133 320           318          70
#> 134 335            40         331
#> 135 332            79         331
#> 136 333           331          80
#> 137 328           318         321
#> 138 329           315         324
#> 139 373           368         367
#> 140 372           366         367
#> 141 375           365         370
#> 142 374           369         368
#> 143 371           365         366
#> 144 381           359         360
#> 145 380           361         364
#> 146 383           356         355
#> 147 377           363         362
#> 148 376           370         369
#> 149 378           363         364
#> 150 406           402         403
#> 151 407           403         400
#> 152 404           400         401
#> 153 399           393         392
#> 154 398           394         393
#> 155 396           392         395
#> 156 387           353         352
#> 157 386           358         359
#> 158 385           357         358
#> 159 384           355         357
#> 160 391           354         353
#> 161 390           350         354
#> 162 389           351         350
#> 163 388           352         351
#> 164 508           483         482
#> 165 509           492         493
#> 166 510           493         490
#> 167 511           490         491
#> 168 504           481         478
#> 169 505           482         485
#> 170 507           484         483
#> 171 501           477         476
#> 172 502           477         480
#> 173 496           474         475
#> 174 498           476         479
#> 175 499           479         478
#> 176 495           473         474
#> 177 494           472         473
#> 178 550           431         430
#> 179 551           430         429
#> 180 548           433         432
#> 181 546           435         434
#> 182 547           434         433
#> 183 544           437         436
#> 184 558           423         422
#> 185 559           422         421
#> 186 556           425         424
#> 187 554           427         426
#> 188 555           426         425
#> 189 552           429         428
#> 190 567           415         414
#> 191 566           413         415
#> 192 564           416         419
#> 193 563           417         416
#> 194 562           418         417
#> 195 560           421         420
#> 196 575           466         467
#> 197 573           409         408
#> 198 572           410         409
#> 199 571           411         410
#> 200 568           414         412
#> 201 518           469         470
#> 202 519           470         471
#> 203 513           488         489
#> 204 514           489         486
#> 205 515           486         487
#> 206 524           456         457
#> 207 525           457         458
#> 208 526           454         453
#> 209 527           453         452
#> 210 520           471         468
#> 211 521           458         459
#> 212 523           460         456
#> 213 532           449         448
#> 214 535           446         445
#> 215 534           447         446
#> 216 529           455         454
#> 217 531           450         449
#> 218 530           450         451
#> 219 540           441         440
#> 220 543           438         437
#> 221 542           439         438
#> 222 536           445         444
#> 223 539           442         441
#> 224 538           443         442
#> 225 610           606         609
#> 226 611           609         356
#> 227 614           612          20
#> 228 619            49         618
#> 229 616           594         615
#> 230 617           615         612
#> 231 622           401         621
#> 232 623           621         402
#> 233 620           618          48
#> 234 626           624          22
#> 235 631           475         630
#> 236 629           627         624
#> 237 628            48         627
#> 238 632           630         472
#> 239 638           636         633
#> 240 637            39         636
#> 241 576           467         462
#> 242 577           462         463
#> 243 578           463         464
#> 244 579           464         461
#> 245 580           461         465
#> 246 583             4         582
#> 247 587           585           4
#> 248 589           366         582
#> 249 590           585         367
#> 250 592           582         591
#> 251 595           591         594
#> 252 598           361         597
#> 253 601           597         600
#> 254 602           600         362
#> 255 605           591         600
#> 256 604           597         594
#> 257 607           360         606
#> 258 686           684         681
#> 259 685            30         684
#> 260 683           681          29
#> 261 680           678         460
#> 262 679           675         678
#> 263 676           459         675
#> 264 674           672         669
#> 265 673           663         672
#> 266 700           448         699
#> 267 701           699         451
#> 268 698           696         693
#> 269 697            29         696
#> 270 692           690         687
#> 271 691           452         690
#> 272 689           687         455
#> 273 653           651          16
#> 274 652            19         651
#> 275 655           491         654
#> 276 649           485         648
#> 277 650           648         484
#> 278 644           642          41
#> 279 647           645         477
#> 280 646           478         645
#> 281 641           639         481
#> 282 640           480         639
#> 283 643            50         642
#> 284 668           666         488
#> 285 671           669          33
#> 286 664           660         663
#> 287 667           487         666
#> 288 661           633         660
#> 289 658           654         657
#> 290 659           657         492
#> 291 745           738         744
#> 292 751           428         750
#> 293 748           744         747
#> 294 739           735         738
#> 295 736           732         735
#> 296 743           741         435
#> 297 742           432         741
#> 298 763           759         762
#> 299 760           756         759
#> 300 766           762         765
#> 301 754           424         753
#> 302 755           753         427
#> 303 752           750         431
#> 304 757           747         756
#> 305 712           702         711
#> 306 715           711         714
#> 307 718           714         717
#> 308 707           705         702
#> 309 706           693         705
#> 310 709           444         708
#> 311 710           708         447
#> 312 728           726         439
#> 313 730           720         729
#> 314 733           729         732
#> 315 721           717         720
#> 316 724           440         723
#> 317 725           723         443
#> 318 727           436         726
#> 319 821           819           7
#> 320 823             0         822
#> 321 817           465         816
#> 322 818           816         813
#> 323 830           828          31
#> 324 824           822         819
#> 325 827           825         469
#> 326 826           468         825
#> 327 805           395         804
#> 328 806           804         801
#> 329 800           798          28
#> 330 803           801         394
#> 331 812           810         807
#> 332 815           813         466
#> 333 811             2         810
#> 334 790           408         789
#> 335 788           786         413
#> 336 787           412         786
#> 337 784           780         783
#> 338 799           795         798
#> 339 796           783         795
#> 340 794           792         411
#> 341 793           789         792
#> 342 775           765         774
#> 343 772           419         771
#> 344 773           771         418
#> 345 770           768         423
#> 346 769           420         768
#> 347 781           777         780
#> 348 778           774         777
#> 349 880           666         669
#> 350 885           807         884
#> 351 886           884         881
#> 352 889           887          11
#> 353 888            13         887
#> 354 891            38         890
#> 355 892           890          13
#> 356 895           884         887
#> 357 894           881         890
#> 358 864           804         807
#> 359 865           606         615
#> 360 866           612         609
#> 361 868           351          49
#> 362 869           621         624
#> 363 870           627         401
#> 364 871            50         639
#> 365 872           642         480
#> 366 873           630         633
#> 367 874           472         636
#> 368 875           648          19
#> 369 876           484         651
#> 370 877           657         660
#> 371 878           654         663
#> 372 879           488         672
#> 373 851           744         431
#> 374 850           741         738
#> 375 849           735         435
#> 376 848           436         732
#> 377 855           762         423
#> 378 854           424         759
#> 379 853           756         753
#> 380 852           750         747
#> 381 859           780         413
#> 382 858           419         777
#> 383 857           774         771
#> 384 856           768         765
#> 385 863           810         801
#> 386 862           798         789
#> 387 861           795         792
#> 388 860           786         783
#> 389 835           825         828
#> 390 832            17         831
#> 391 833           831         828
#> 392 838           684         675
#> 393 839           696         687
#> 394 836           468         831
#> 395 837           678         681
#> 396 842           699         702
#> 397 843           711         708
#> 398 840           690         693
#> 399 841           705         451
#> 400 846           723         720
#> 401 847           729         726
#> 402 844           444         714
#> 403 845           717         443
#> 404 931           813         822
#> 405 930           819         816
#> 406 906           618         352
#> 407 973            92         714
#> 408 972            88         705
#> 409 962            73         881
#> 410 961            73         884
#> 411 967            72         963
#> 412 965           963         585
#> 413 964           881         963
#> 
#> $raw_parsed_data$faces
#>      id     type orientation edge_hrefs neighbour
#> 1   956 building           -        772       927
#> 2   956 building           -        564      <NA>
#> 3   956 building           -        563      <NA>
#> 4   956 building           -        562      <NA>
#> 5   956 building           -        773      <NA>
#> 6   957 building           +        788       928
#> 7   957 building           +        566      <NA>
#> 8   957 building           +        567      <NA>
#> 9   957 building           +        568      <NA>
#> 10  957 building           +        787      <NA>
#> 11  958 building           +        794      <NA>
#> 12  958 building           +        571      <NA>
#> 13  958 building           +        572      <NA>
#> 14  958 building           +        573      <NA>
#> 15  958 building           +        790      <NA>
#> 16  958 building           +        793       929
#> 17  959 building           +        806       899
#> 18  959 building           +        803      <NA>
#> 19  959 building           +        398      <NA>
#> 20  959 building           +        399      <NA>
#> 21  959 building           +        396      <NA>
#> 22  959 building           +        805      <NA>
#> 23  952 building           +        550      <NA>
#> 24  952 building           +        551      <NA>
#> 25  952 building           +        552      <NA>
#> 26  952 building           +        751      <NA>
#> 27  952 building           +        752       924
#> 28  953 building           +        575      <NA>
#> 29  953 building           +        576      <NA>
#> 30  953 building           +        577      <NA>
#> 31  953 building           +        578      <NA>
#> 32  953 building           +        579      <NA>
#> 33  953 building           +        580      <NA>
#> 34  953 building           +        817      <NA>
#> 35  953 building           +        818       932
#> 36  953 building           +        815      <NA>
#> 37  954 building           -        754       925
#> 38  954 building           -        556      <NA>
#> 39  954 building           -        555      <NA>
#> 40  954 building           -        554      <NA>
#> 41  954 building           -        755      <NA>
#> 42  955 building           +        770       926
#> 43  955 building           +        558      <NA>
#> 44  955 building           +        559      <NA>
#> 45  955 building           +        560      <NA>
#> 46  955 building           +        769      <NA>
#> 47  948 building           +        534      <NA>
#> 48  948 building           +        535      <NA>
#> 49  948 building           +        536      <NA>
#> 50  948 building           +        709       920
#> 51  948 building           +        710      <NA>
#> 52  949 building           +        538      <NA>
#> 53  949 building           +        539      <NA>
#> 54  949 building           +        540      <NA>
#> 55  949 building           +        724      <NA>
#> 56  949 building           +        725       921
#> 57  950 building           +        542      <NA>
#> 58  950 building           +        543      <NA>
#> 59  950 building           +        544      <NA>
#> 60  950 building           +        727       922
#> 61  950 building           +        728      <NA>
#> 62  951 building           -        546      <NA>
#> 63  951 building           -        743       923
#> 64  951 building           -        742      <NA>
#> 65  951 building           -        548      <NA>
#> 66  951 building           -        547      <NA>
#> 67  944 building           -        827      <NA>
#> 68  944 building           -        826       916
#> 69  944 building           -        520      <NA>
#> 70  944 building           -        519      <NA>
#> 71  944 building           -        518      <NA>
#> 72  945 building           -        676      <NA>
#> 73  945 building           -        521      <NA>
#> 74  945 building           -        525      <NA>
#> 75  945 building           -        524      <NA>
#> 76  945 building           -        523      <NA>
#> 77  945 building           -        680      <NA>
#> 78  945 building           -        679       917
#> 79  946 building           +        529      <NA>
#> 80  946 building           +        526      <NA>
#> 81  946 building           +        527      <NA>
#> 82  946 building           +        691      <NA>
#> 83  946 building           +        692       918
#> 84  946 building           +        689      <NA>
#> 85  947 building           -        530      <NA>
#> 86  947 building           +        531      <NA>
#> 87  947 building           +        532      <NA>
#> 88  947 building           +        700      <NA>
#> 89  947 building           +        701       919
#> 90  941 building           -        508      <NA>
#> 91  941 building           -        507      <NA>
#> 92  941 building           -        650       913
#> 93  941 building           -        649      <NA>
#> 94  941 building           -        505      <NA>
#> 95  940 building           +        504      <NA>
#> 96  940 building           +        646      <NA>
#> 97  940 building           +        647       939
#> 98  940 building           +        502      <NA>
#> 99  940 building           +        640       912
#> 100 940 building           +        641      <NA>
#> 101 943 building           +        513      <NA>
#> 102 943 building           +        514      <NA>
#> 103 943 building           +        515      <NA>
#> 104 943 building           +        667      <NA>
#> 105 943 building           +        668       915
#> 106 942 building           +        659      <NA>
#> 107 942 building           +        509      <NA>
#> 108 942 building           +        510      <NA>
#> 109 942 building           +        511      <NA>
#> 110 942 building           +        655      <NA>
#> 111 942 building           +        658       914
#> 112 937 building           -        406      <NA>
#> 113 937 building           -        623      <NA>
#> 114 937 building           -        622       933
#> 115 937 building           -        404      <NA>
#> 116 937 building           -        407      <NA>
#> 117 936 building           -        386      <NA>
#> 118 936 building           -        385      <NA>
#> 119 936 building           -        384      <NA>
#> 120 936 building           -        383      <NA>
#> 121 936 building           -        611      <NA>
#> 122 936 building           -        610       904
#> 123 936 building           -        607      <NA>
#> 124 936 building           -        381      <NA>
#> 125 939 building           -        498      <NA>
#> 126 939 building           -        501      <NA>
#> 127 939 building           -        647       940
#> 128 939 building           -        646      <NA>
#> 129 939 building           -        499      <NA>
#> 130 938 building           +        495      <NA>
#> 131 938 building           +        496      <NA>
#> 132 938 building           +        631      <NA>
#> 133 938 building           +        632       911
#> 134 938 building           +        494      <NA>
#> 135 935 building           -        377      <NA>
#> 136 935 building           +        378      <NA>
#> 137 935 building           -        380      <NA>
#> 138 935 building           +        598      <NA>
#> 139 935 building           +        601       903
#> 140 935 building           +        602      <NA>
#> 141 934 building           -        373      <NA>
#> 142 934 building           -        374      <NA>
#> 143 934 building           -        376      <NA>
#> 144 934 building           -        375      <NA>
#> 145 934 building           +        371      <NA>
#> 146 934 building           +        372       902
#> 147 254 building           +        207      <NA>
#> 148 254 building           +        208      <NA>
#> 149 254 building           +        126      <NA>
#> 150 254 building           +        123       974
#> 151 254 building           +        223      <NA>
#> 152 254 building           +        228      <NA>
#> 153 254 building           +        227      <NA>
#> 154 254 building           +        209      <NA>
#> 155 255 building           +        185      <NA>
#> 156 255 building           +        127      <NA>
#> 157 255 building           +        128       976
#> 158 255 building           +        229       274
#> 159 255 building           +        232      <NA>
#> 160 255 building           +        178      <NA>
#> 161 255 building           +        177      <NA>
#> 162 253 building           +        234       297
#> 163 253 building           +        115      <NA>
#> 164 253 building           +        200      <NA>
#> 165 253 building           +        196      <NA>
#> 166 253 building           +        244      <NA>
#> 167 253 building           +        241       297
#> 168 250 building           +        332       336
#> 169 250 building           +        333      <NA>
#> 170 250 building           +        203      <NA>
#> 171 250 building           +        212      <NA>
#> 172 250 building           +        213      <NA>
#> 173 250 building           +        211      <NA>
#> 174 250 building           +        210      <NA>
#> 175 251 building           +        197      <NA>
#> 176 251 building           +        120      <NA>
#> 177 251 building           +        122      <NA>
#> 178 251 building           +        205       968
#> 179 251 building           +        221      <NA>
#> 180 251 building           +        222      <NA>
#> 181 905 building           +        388       910
#> 182 905 building           +        389      <NA>
#> 183 905 building           +        390      <NA>
#> 184 905 building           +        391      <NA>
#> 185 905 building           +        387      <NA>
#> 186 248 building           +        190      <NA>
#> 187 248 building           +        191      <NA>
#> 188 248 building           +        199      <NA>
#> 189 248 building           +        338      <NA>
#> 190 248 building           +        339       349
#> 191 249 building           +        186      <NA>
#> 192 249 building           +        187      <NA>
#> 193 249 building           +        300      <NA>
#> 194 249 building           +        303       314
#> 195 249 building           +        304      <NA>
#> 196 249 building           +        182      <NA>
#> 197 249 building           +        183      <NA>
#> 198 249 building           +        184      <NA>
#> 199 249 building           +        192      <NA>
#> 200 249 building           +        193      <NA>
#> 201 247 building           +        194      <NA>
#> 202 247 building           +        316      <NA>
#> 203 247 building           +        317       330
#> 204 247 building           +        319       330
#> 205 247 building           +        320      <NA>
#> 206 247 building           +        189      <NA>
#> 207 298 building           +        290      <NA>
#> 208 298 building           -        206      <NA>
#> 209 298 building           -        202      <NA>
#> 210 298 building           -        201      <NA>
#> 211 298 building           +        288      <NA>
#> 212 298 building           +        289       297
#> 213 960 building           -        888       896
#> 214 960 building           -        892       896
#> 215 960 building           -        891      <NA>
#> 216 960 building           -        150      <NA>
#> 217 960 building           -        225      <NA>
#> 218 960 building           -        245      <NA>
#> 219 960 building           -        163      <NA>
#> 220 960 building           -        164      <NA>
#> 221 960 building           -        159      <NA>
#> 222 960 building           -        158      <NA>
#> 223 960 building           -        220      <NA>
#> 224 960 building           -        217      <NA>
#> 225 960 building           -        121      <NA>
#> 226 960 building           -        889      <NA>
#> 227 275     road           +        136       268
#> 228 275     road           +        139      <NA>
#> 229 275     road           +        138       258
#> 230 275     road           +        141      <NA>
#> 231 274     road           +        216      <NA>
#> 232 274     road           -        229       255
#> 233 274     road           +        230       976
#> 234 274     road           +        142      <NA>
#> 235 274     road           +        140       279
#> 236 273     road           +        134       281
#> 237 273     road           +        306      <NA>
#> 238 273     road           +        309       314
#> 239 273     road           +        322      <NA>
#> 240 273     road           +        325       330
#> 241 273     road           +        326      <NA>
#> 242 273     road           -        109       969
#> 243 273     road           +        583       902
#> 244 273     road           +        592      <NA>
#> 245 273     road           +        595       903
#> 246 273     road           +        616      <NA>
#> 247 273     road           +        617       904
#> 248 273     road           +        614      <NA>
#> 249 279     road           -        140       274
#> 250 279     road           -        107       266
#> 251 279     road           -        110       971
#> 252 278     road           -        173       271
#> 253 278     road           -        170       268
#> 254 278     road           -        130       263
#> 255 281     road           -        171       256
#> 256 281     road           -        157       296
#> 257 281     road           -        134       273
#> 258 280     road           -        148       257
#> 259 280     road           -        165       262
#> 260 280     road           -        154       259
#> 261 258     road           -        138       275
#> 262 258     road           +        160      <NA>
#> 263 258     road           -        112       269
#> 264 258     road           +        162      <NA>
#> 265 259     road           +        154       280
#> 266 259     road           +        240       336
#> 267 259     road           +        237      <NA>
#> 268 259     road           -        146       297
#> 269 259     road           +        161      <NA>
#> 270 256     road           +        171       281
#> 271 256     road           +        180      <NA>
#> 272 256     road           -        175       271
#> 273 256     road           +        643       912
#> 274 256     road           +        644      <NA>
#> 275 257     road           +        145       270
#> 276 257     road           +        147      <NA>
#> 277 257     road           +        148       280
#> 278 257     road           +        149      <NA>
#> 279 262     road           -        144       974
#> 280 262     road           +        235      <NA>
#> 281 262     road           +        236      <NA>
#> 282 262     road           +        165       280
#> 283 262     road           +        685      <NA>
#> 284 262     road           +        686       917
#> 285 262     road           +        683      <NA>
#> 286 263     road           +        130       278
#> 287 263     road           +        133      <NA>
#> 288 263     road           +        131       270
#> 289 263     road           +        652       913
#> 290 263     road           +        653      <NA>
#> 291 266     road           +        103       269
#> 292 266     road           +        106      <NA>
#> 293 266     road           +        107       279
#> 294 266     road           +        108      <NA>
#> 295 933     road           +        869      <NA>
#> 296 933     road           -        629       268
#> 297 933     road           +        870      <NA>
#> 298 933     road           +        622       937
#> 299 932     road           +        818       953
#> 300 932     road           +        931      <NA>
#> 301 932     road           +        824       269
#> 302 932     road           +        930      <NA>
#> 303 270     road           -        131       263
#> 304 270     road           +        832      <NA>
#> 305 270     road           +        833       916
#> 306 270     road           +        830      <NA>
#> 307 270     road           -        145       257
#> 308 270     road           +        167      <NA>
#> 309 929     road           +        861      <NA>
#> 310 929     road           -        793       958
#> 311 929     road           -        862      <NA>
#> 312 929     road           -        799       976
#> 313 271     road           +        175       256
#> 314 271     road           +        619       910
#> 315 271     road           +        620      <NA>
#> 316 271     road           +        173       278
#> 317 271     road           +        172      <NA>
#> 318 928     road           -        788       957
#> 319 928     road           +        860      <NA>
#> 320 928     road           -        784       976
#> 321 928     road           +        859      <NA>
#> 322 268     road           +        170       278
#> 323 268     road           +        628      <NA>
#> 324 268     road           +        629       933
#> 325 268     road           +        626      <NA>
#> 326 268     road           -        136       275
#> 327 268     road           +        169      <NA>
#> 328 269     road           +        112       258
#> 329 269     road           +        116      <NA>
#> 330 269     road           -        103       266
#> 331 269     road           +        823      <NA>
#> 332 269     road           +        824       932
#> 333 269     road           +        821      <NA>
#> 334 926     road           +        855      <NA>
#> 335 926     road           -        770       955
#> 336 926     road           +        856      <NA>
#> 337 926     road           -        766       976
#> 338 927     road           +        857      <NA>
#> 339 927     road           -        772       956
#> 340 927     road           +        858      <NA>
#> 341 927     road           -        778       976
#> 342 924     road           +        851      <NA>
#> 343 924     road           -        752       952
#> 344 924     road           +        852      <NA>
#> 345 924     road           -        748       976
#> 346 925     road           +        853      <NA>
#> 347 925     road           -        754       954
#> 348 925     road           +        854      <NA>
#> 349 925     road           -        760       976
#> 350 922     road           +        847      <NA>
#> 351 922     road           -        727       950
#> 352 922     road           +        848      <NA>
#> 353 922     road           -        733       976
#> 354 923     road           +        849      <NA>
#> 355 923     road           -        743       951
#> 356 923     road           +        850      <NA>
#> 357 923     road           -        739       976
#> 358 920     road           -        709       948
#> 359 920     road           +        844      <NA>
#> 360 920     road           -        715       975
#> 361 920     road           +        843      <NA>
#> 362 921     road           +        845      <NA>
#> 363 921     road           -        725       949
#> 364 921     road           +        846      <NA>
#> 365 921     road           -        721       976
#> 366 918     road           +        839      <NA>
#> 367 918     road           -        692       946
#> 368 918     road           +        840      <NA>
#> 369 918     road           -        698       974
#> 370 919     road           +        841      <NA>
#> 371 919     road           -        701       947
#> 372 919     road           +        842      <NA>
#> 373 919     road           -        707       975
#> 374 916     road           +        826       944
#> 375 916     road           +        835      <NA>
#> 376 916     road           -        833       270
#> 377 916     road           -        836      <NA>
#> 378 314     road           -        312      <NA>
#> 379 314     road           -        303       249
#> 380 314     road           +        313      <NA>
#> 381 314     road           -        309       273
#> 382 917     road           +        837      <NA>
#> 383 917     road           -        686       262
#> 384 917     road           +        838      <NA>
#> 385 917     road           +        679       945
#> 386 914     road           -        658       942
#> 387 914     road           +        878      <NA>
#> 388 914     road           -        664       297
#> 389 914     road           -        877      <NA>
#> 390 915     road           -        668       943
#> 391 915     road           +        880      <NA>
#> 392 915     road           -        674       297
#> 393 915     road           -        879      <NA>
#> 394 912     road           -        640       940
#> 395 912     road           -        872      <NA>
#> 396 912     road           -        643       256
#> 397 912     road           +        871      <NA>
#> 398 913     road           -        875      <NA>
#> 399 913     road           +        650       941
#> 400 913     road           +        876      <NA>
#> 401 913     road           -        652       263
#> 402 911     road           +        632       938
#> 403 911     road           +        874      <NA>
#> 404 911     road           +        638       297
#> 405 911     road           -        873      <NA>
#> 406 910     road           +        388       905
#> 407 910     road           +        868      <NA>
#> 408 910     road           +        619       271
#> 409 910     road           +        906      <NA>
#> 410 904     road           -        610       936
#> 411 904     road           +        865      <NA>
#> 412 904     road           +        617       273
#> 413 904     road           +        866      <NA>
#> 414 903     road           -        595       273
#> 415 903     road           +        605      <NA>
#> 416 903     road           -        601       935
#> 417 903     road           +        604      <NA>
#> 418 296     road           +        153      <NA>
#> 419 296     road           +        294       297
#> 420 296     road           -        292      <NA>
#> 421 296     road           +        157       281
#> 422 297     road           -        294       296
#> 423 297     road           +        637      <NA>
#> 424 297     road           +        638       911
#> 425 297     road           +        661      <NA>
#> 426 297     road           +        664       914
#> 427 297     road           +        673      <NA>
#> 428 297     road           +        674       915
#> 429 297     road           +        671      <NA>
#> 430 297     road           +        146       259
#> 431 297     road           +        117      <NA>
#> 432 297     road           -        234       253
#> 433 297     road           -        241       253
#> 434 297     road           +        293      <NA>
#> 435 297     road           +        289       298
#> 436 902     road           -        372       934
#> 437 902     road           +        589      <NA>
#> 438 902     road           -        583       273
#> 439 902     road           -        587       969
#> 440 902     road           +        590      <NA>
#> 441 899     road           -        806       959
#> 442 899     road           +        864      <NA>
#> 443 899     road           -        812       971
#> 444 899     road           +        863      <NA>
#> 445 896     road           -        894      <NA>
#> 446 896     road           -        886       970
#> 447 896     road           +        895      <NA>
#> 448 896     road           -        888       960
#> 449 896     road           -        892       960
#> 450 336     road           -        332       250
#> 451 336     road           -        240       259
#> 452 336     road           +        335      <NA>
#> 453 349     road           +        348      <NA>
#> 454 349     road           -        345       969
#> 455 349     road           +        347      <NA>
#> 456 349     road           -        339       248
#> 457 330     road           -        328      <NA>
#> 458 330     road           -        319       247
#> 459 330     road           -        317       247
#> 460 330     road           +        329      <NA>
#> 461 330     road           -        325       273
#> 462 976     road           -        973       975
#> 463 976     road           +        228      <NA>
#> 464 976     road           -        104      <NA>
#> 465 976     road           +        128       255
#> 466 976     road           +        230       274
#> 467 976     road           -        800      <NA>
#> 468 976     road           -        799       929
#> 469 976     road           -        796      <NA>
#> 470 976     road           -        784       928
#> 471 976     road           -        781      <NA>
#> 472 976     road           -        778       927
#> 473 976     road           -        775      <NA>
#> 474 976     road           -        766       926
#> 475 976     road           -        763      <NA>
#> 476 976     road           -        760       925
#> 477 976     road           -        757      <NA>
#> 478 976     road           -        748       924
#> 479 976     road           -        745      <NA>
#> 480 976     road           -        739       923
#> 481 976     road           -        736      <NA>
#> 482 976     road           -        733       922
#> 483 976     road           -        730      <NA>
#> 484 976     road           -        721       921
#> 485 976     road           -        718      <NA>
#> 486 971     road           +        885      <NA>
#> 487 971     road           -        961       970
#> 488 971     road           -        122      <NA>
#> 489 971     road           +        119      <NA>
#> 490 971     road           +        110       279
#> 491 971     road           +        811      <NA>
#> 492 971     road           +        812       899
#> 493 970     road           +        961       971
#> 494 970     road           +        886       896
#> 495 970     road           -        962       968
#> 496 969     road           +        965      <NA>
#> 497 969     road           +        587       902
#> 498 969     road           +        109       273
#> 499 969     road           +        344      <NA>
#> 500 969     road           +        345       349
#> 501 969     road           +        342      <NA>
#> 502 969     road           -        221      <NA>
#> 503 969     road           +        967       968
#> 504 968     road           -        205       251
#> 505 968     road           +        962       970
#> 506 968     road           +        964      <NA>
#> 507 968     road           -        967       969
#> 508 975     road           +        972       974
#> 509 975     road           +        707       919
#> 510 975     road           +        712      <NA>
#> 511 975     road           +        715       920
#> 512 975     road           -        973       976
#> 513 975     road           -        223      <NA>
#> 514 974     road           +        972       975
#> 515 974     road           -        706      <NA>
#> 516 974     road           -        698       918
#> 517 974     road           -        697      <NA>
#> 518 974     road           -        144       262
#> 519 974     road           -        129      <NA>
#> 520 974     road           +        123       254
#> 
#> 
#> attr(,"class")
#> [1] "rrs_map" "list"